requirements.txt
cobs==1.2.1
crccheck==1.3.0
pyserial==3.5
